# Pass ip, ips, and hostname to Request object