--- a/fastify.d.ts
+++ b/fastify.d.ts
@@ -193,7 +193,7 @@ declare namespace fastify {
       },
       deriveVersion<Context>(req: Object, ctx?: Context) : String,
     },
-    modifyCoreObjects: boolean
+    modifyCoreObjects?: boolean
   }
   interface ServerOptionsAsSecure extends ServerOptions {
     https: http2.SecureServerOptions