--- a/lib/reply.js
+++ b/lib/reply.js
@@ -90,7 +90,7 @@ Reply.prototype.send = function (payload) {
   }
 
   if (this[kReplySent]) {
-    this.res.log.warn({ err: new FST_ERR_REP_ALREADY_SENT() }, 'Reply already sent')
+    this.log.warn({ err: new FST_ERR_REP_ALREADY_SENT() }, 'Reply already sent')
     return
   }
 
@@ -318,7 +318,7 @@ function sendStream (payload, res, reply) {
     sourceOpen = false
     if (err != null) {
       if (res.headersSent) {
-        res.log.warn({ err }, 'response terminated with an error with headers already sent')
+        reply.log.warn({ err }, 'response terminated with an error with headers already sent')
         res.destroy()
       } else {
         onErrorHook(reply, err)
@@ -330,7 +330,7 @@ function sendStream (payload, res, reply) {
   eos(res, function (err) {
     if (err != null) {
       if (res.headersSent) {
-        res.log.warn({ err }, 'response terminated with an error with headers already sent')
+        reply.log.warn({ err }, 'response terminated with an error with headers already sent')
       }
       if (sourceOpen) {
         if (payload.destroy) {
@@ -353,7 +353,7 @@ function sendStream (payload, res, reply) {
       res.setHeader(key, reply[kReplyHeaders][key])
     }
   } else {
-    res.log.warn('response will send, but you shouldn\'t use res.writeHead in stream mode')
+    reply.log.warn('response will send, but you shouldn\'t use res.writeHead in stream mode')
   }
   payload.pipe(res)
 }
@@ -500,7 +500,7 @@ function notFound (reply) {
   reply[kReplyIsError] = false
 
   if (reply.context[kFourOhFourContext] === null) {
-    reply.res.log.warn('Trying to send a NotFound error inside a 404 handler. Sending basic 404 response.')
+    reply.log.warn('Trying to send a NotFound error inside a 404 handler. Sending basic 404 response.')
     reply.code(404).send('404 Not Found')
     return
   }