--- a/lib/request.js
+++ b/lib/request.js
@@ -1,22 +1,28 @@
 'use strict'
 
-function Request (params, req, query, headers, log) {
+function Request (params, req, query, headers, log, ip, ips, hostname) {
   this.params = params
   this.raw = req
   this.query = query
   this.headers = headers
   this.log = log
   this.body = null
+  this.ip = ip
+  this.ips = ips
+  this.hostname = hostname
 }
 
 function buildRequest (R) {
-  function _Request (params, req, query, headers, log) {
+  function _Request (params, req, query, headers, log, ip, ips, hostname) {
     this.params = params
     this.raw = req
     this.query = query
     this.headers = headers
     this.log = log
     this.body = null
+    this.ip = ip
+    this.ips = ips
+    this.hostname = hostname
   }
   _Request.prototype = new R()