--- a/lib/wrapThenable.js
+++ b/lib/wrapThenable.js
@@ -31,11 +31,11 @@ function wrapThenable (thenable, reply) {
         reply.send(err)
       }
     } else if (reply[kReplySent] === false) {
-      reply.res.log.error({ err: new FST_ERR_PROMISE_NOT_FULLFILLED() }, `Promise may not be fulfilled with 'undefined' when statusCode is not 204`)
+      reply.log.error({ err: new FST_ERR_PROMISE_NOT_FULLFILLED() }, `Promise may not be fulfilled with 'undefined' when statusCode is not 204`)
     }
   }, function (err) {
     if (reply[kReplySentOverwritten] === true) {
-      reply.res.log.error({ err }, 'Promise errored, but reply.sent = true was set')
+      reply.log.error({ err }, 'Promise errored, but reply.sent = true was set')
       return
     }
     reply[kReplySent] = false