--- a/test/trust-proxy.test.js
+++ b/test/trust-proxy.test.js
@@ -18,20 +18,25 @@ const sgetForwardedRequest = (app, forHeader, path) => {
 
 const testRequestValues = (t, req, options) => {
   if (options.ip) {
+    t.ok(req.raw.ip, 'ip is defined')
+    t.equal(req.raw.ip, options.ip, 'gets ip from x-forwarder-for')
     t.ok(req.ip, 'ip is defined')
     t.equal(req.ip, options.ip, 'gets ip from x-forwarder-for')
   }
   if (options.hostname) {
+    t.ok(req.raw.hostname, 'hostname is defined')
+    t.equal(req.raw.hostname, options.hostname, 'gets hostname from x-forwarded-host')
     t.ok(req.hostname, 'hostname is defined')
     t.equal(req.hostname, options.hostname, 'gets hostname from x-forwarded-host')
   }
   if (options.ips) {
+    t.deepEqual(req.raw.ips, options.ips, 'gets ips from x-forwarder-for')
     t.deepEqual(req.ips, options.ips, 'gets ips from x-forwarder-for')
   }
 }
 
 test('trust proxy', (t) => {
-  t.plan(8)
+  t.plan(15)
   const app = fastify({
     trustProxy: true
   })
@@ -56,7 +61,7 @@ test('trust proxy', (t) => {
 })
 
 test('trust proxy chain', (t) => {
-  t.plan(3)
+  t.plan(5)
   const app = fastify({
     trustProxy: ['127.0.0.1', '192.168.1.1']
   })
@@ -76,7 +81,7 @@ test('trust proxy chain', (t) => {
 })
 
 test('trust proxy function', (t) => {
-  t.plan(3)
+  t.plan(5)
   const app = fastify({
     trustProxy: (address) => address === '127.0.0.1'
   })
@@ -95,7 +100,7 @@ test('trust proxy function', (t) => {
 })
 
 test('trust proxy number', (t) => {
-  t.plan(4)
+  t.plan(7)
   const app = fastify({
     trustProxy: 1
   })
@@ -114,7 +119,7 @@ test('trust proxy number', (t) => {
 })
 
 test('trust proxy IP addresses', (t) => {
-  t.plan(4)
+  t.plan(7)
   const app = fastify({
     trustProxy: '127.0.0.1, 2.2.2.2'
   })