--- a/test/types/index.ts
+++ b/test/types/index.ts
@@ -61,7 +61,8 @@ const cors = require('cors')
     ignoreTrailingSlash: true,
     bodyLimit: 1000,
     maxParamLength: 200,
-    querystringParser: (str: string) => ({ str: str, strArray: [str] })
+    querystringParser: (str: string) => ({ str: str, strArray: [str] }),
+    modifyCoreObjects: true
   })
 
   // custom types