#### Checklist

- [x] run `npm run test` and `npm run benchmark`
- [x] tests and/or benchmarks are included
- [x] documentation is changed or added
- [x] commit message and code follows [Code of conduct](https://github.com/fastify/fastify/blob/master/CODE_OF_CONDUCT.md)

Add ip, ips, and hostname in `Request` object. Related to @delvedor's feedback at https://github.com/fastify/fastify/pull/1086#discussion_r211718687. All credits go to @nwoltman . Without his guidance, the work won't be possible. 

[benchmarks](https://gist.github.com/shwei/390c850ebc7a52650de5e33481f5fce3#file-fastify-3dc46f3-tables-comparing-master-vs-pass-ip-to-request-constructor-branches-L1)