@shwei 2019-02-26
> Is this a breaking change?
> We are no longer writing the values in the Node Core request object, but in ours.

Hi @delvedor: Do you mean this particular commit a breaking change or this pull request?  

The call to `handleTrustProxy(req)` assigns req.ip to some value depending on the proxy option. Isn't the req object in this case a Node Core one? https://github.com/fastify/fastify/blob/master/fastify.js#L303