@nwoltman 2019-02-27
@delvedor 

> Is this a breaking change?
> We are no longer writing the values in the Node Core request object, but in ours

Yes, removing those values makes this a breaking change. Another option would be to add the values to Fastify's `Request` object now, and remove them from the Node core `req` object for the next major.