@shwei 2019-02-28
master branch:
```>$ git checkout master
Switched to branch 'master'
Your branch is ahead of 'origin/master' by 6 commits.
  (use "git push" to publish your local commits)
>$ node examples/example.js 
server listening on 3000
>$ autocannon -c 100 -d 40 -p 10 localhost:3000
Running 40s test @ http://localhost:3000
100 connections with 10 pipelining factor

┌─────────┬──────┬──────┬───────┬───────┬─────────┬──────────┬───────────┐
│ Stat    │ 2.5% │ 50%  │ 97.5% │ 99%   │ Avg     │ Stdev    │ Max       │
├─────────┼──────┼──────┼───────┼───────┼─────────┼──────────┼───────────┤
│ Latency │ 0 ms │ 0 ms │ 47 ms │ 54 ms │ 4.31 ms │ 13.76 ms │ 400.59 ms │
└─────────┴──────┴──────┴───────┴───────┴─────────┴──────────┴───────────┘
┌───────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┐
│ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%   │ Avg     │ Stdev   │ Min     │
├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
│ Req/Sec   │ 11895   │ 11895   │ 22175   │ 26943   │ 22765.1 │ 3482.72 │ 11890   │
├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
│ Bytes/Sec │ 1.95 MB │ 1.95 MB │ 3.64 MB │ 4.42 MB │ 3.73 MB │ 571 kB  │ 1.95 MB │
└───────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┘

Req/Bytes counts sampled once per second.

911k requests in 40.07s, 149 MB read
```

pass-ip-to-request-constructor:
```
>$ git checkout pass-ip-to-request-constructor
Switched to branch 'pass-ip-to-request-constructor'
Your branch is ahead of 'origin/pass-ip-to-request-constructor' by 2 commits.
  (use "git push" to publish your local commits)
>$ node examples/example.js 
server listening on 3000
>$ autocannon -c 100 -d 40 -p 10 localhost:3000
Running 40s test @ http://localhost:3000
100 connections with 10 pipelining factor

┌─────────┬──────┬──────┬───────┬───────┬─────────┬──────────┬───────────┐
│ Stat    │ 2.5% │ 50%  │ 97.5% │ 99%   │ Avg     │ Stdev    │ Max       │
├─────────┼──────┼──────┼───────┼───────┼─────────┼──────────┼───────────┤
│ Latency │ 0 ms │ 0 ms │ 47 ms │ 54 ms │ 4.27 ms │ 13.28 ms │ 409.13 ms │
└─────────┴──────┴──────┴───────┴───────┴─────────┴──────────┴───────────┘
┌───────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┐
│ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%   │ Avg     │ Stdev   │ Min     │
├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
│ Req/Sec   │ 12271   │ 12271   │ 23583   │ 26463   │ 23005.3 │ 2918.96 │ 12270   │
├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
│ Bytes/Sec │ 2.01 MB │ 2.01 MB │ 3.87 MB │ 4.34 MB │ 3.77 MB │ 478 kB  │ 2.01 MB │
└───────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┘

Req/Bytes counts sampled once per second.

920k requests in 40.07s, 151 MB read
```

Per @mcollina 's comment:

> Can you run a benchmark, like so:
> 
> ```
> node examples/example
> ```
> ```
> autocannon -c 100 -d 40 -p 10 localhost:3000
> ```
> And post the results, before and after this PR?